requirements.txt
asgiref==3.8.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
fonttools==4.58.1
inflection==0.5.1
iniconfig==2.1.0
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pluggy==1.6.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyparsing==3.2.3
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
